Address review: move §6.9.2 tests to a dedicated file, trim comments, wire reject paths
Moved the two client flow-control tests to node-http2-flow-control.test.ts
so the gate does not run into the pre-existing 10k-request maxSessionMemory
stress test that sits at its 150s debug timeout in node-http2.test.js.
Shared the raw-server frame parser between the two tests and gave every
awaited promise a reject path wired to the session error event.
